Create Ticket Type
Write a programming code to program the ticketing kiosks to accommodate this requirement. A brand-new amusement park called "Fantasy Kingdom" is soon to open in the city and is marketed as having breathtaking attractiveness. The theme park features more than 30 thrilling and crazy rides, and as a unique feature of the park, the park authorities have set up a number of ticketing kiosks at the entrance to make it easier for people to buy tickets for both the entrance and rides. As there are different fares for various age groups, entrance tickets are often issued based on age.



Tickets come in two varieties: child tickets and adult tickets. 

Child tickets are issued if the age provided is under 15, 
adult tickets are supplied if the age is more than or equal to 15

Sample test cases :
Input 1 :
20
Output 1 :
Adult Ticket
Input 2 :
12
Output 2 :
Child Ticket