Create Lucky Winner
Create a program to determine if the last digit of a visitor's ticket is 3 or 8. The "Fantasy Kingdom" amusement park held its opening ceremony, and the park administration offered several lucky prizes to the guests on the first day. The attendees whose ticket numbers have the last digit of 3 or 8 are therefore determined to be the lucky winners, and appealing rewards are waiting to be presented.

Sample test cases :
Input 1 :
43
Output 1 :
Lucky Winner
Input 2 :
41
Output 2 :
Not a Lucky Winner