Create Salary after N days
Write a program to find the salary of a worker for N days. The first-day salary is Rs. 10, the second-day salary is Rs. 20, the third-day salary is Rs. 40, and so on.

Input format :
The input consists of the number of days (N).

Output format :
The output prints the total salary for N days.

Sample test cases :
Input 1 :
1
Output 1 :
10
Input 2 :
3
Output 2 :
70