Create Strong Number
The factorial concept was explained in a math class. A student discovered a connection between the sum of the factorial of digits in a number and the number itself when experimenting with numbers after understanding the idea of factorial. He began looking for additional numbers like this since he was so excited. Write a suitable Java program to check the supplied numbers to aid him.


Explanation:

Input:

145

Output:

145 Strong Number



In this case, the program reads the input integer 145 and calculates the factorial of each digit of the number (1! + 4! + 5!). It then checks if the sum of the factorials is equal to the original number, which is true for 145, so the program prints "145 Strong Number" on the console as the output.

Input format :
The input consists of an integer.

Output format :
The output displays the number followed by Strong Number / Not Strong Number.



Refer to the sample input and output for format specifications.

Sample test cases :
Input 1 :
145
Output 1 :
145 Strong Number
Input 2 :
140
Output 2 :
140 Not Strong Number